% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.R
\name{rc_all}
\alias{rc_all}
\title{Get all countries}
\usage{
rc_all()
}
\value{
tibble: name, official, cca2, cca3, region, subregion, capital,
population, area, currencies, languages, flag
}
\description{
Get all countries
}
