% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.R
\name{rc_code}
\alias{rc_code}
\title{Get country by alpha code}
\usage{
rc_code(code)
}
\arguments{
\item{code}{ISO 3166-1 alpha-2 or alpha-3 code (e.g. "US", "USA", "BR")}
}
\value{
tibble with one row
}
\description{
Get country by alpha code
}
