% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.R
\name{rc_name}
\alias{rc_name}
\title{Search countries by name}
\usage{
rc_name(name)
}
\arguments{
\item{name}{Country name (full or partial, e.g. "united", "brazil")}
}
\value{
tibble of matching countries
}
\description{
Search countries by name
}
