% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.R
\name{rc_region}
\alias{rc_region}
\title{Get countries by region}
\usage{
rc_region(region)
}
\arguments{
\item{region}{Region name: "Africa", "Americas", "Asia", "Europe", "Oceania"}
}
\value{
tibble of countries in that region
}
\description{
Get countries by region
}
